whisper/requirements.txt
yt_dlp
python-slugify

pip==23.0.1
setuptools==67.6.1
wheel==0.40.0
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
# openai-whisper==20230124
#git+https://github.com/openai/whisper.git
# argparse==1.4.0


